requirements.txt
boto3~=1.34.11
PyJWT==1.6.1
pytest~=7.4.4
chalice~=1.30.0
opensearch-py~=2.4.2
apify-client~=1.6.2
langchain~=0.1.1
langchain-openai~=0.0.5
python-dotenv==1.0.0